refactor: extract file tool adapters
